requirements.txt
Flask==3.0.2
flask-cors==4.0.0
PyJWT==2.8.0
flask_sqlalchemy==3.1.1